core/smpp/src/main/java/org/mobicents/smsc/smpp/DbSmsRoutingRule.java
/*
 * TeleStax, Open Source Cloud Communications  Copyright 2012. 
 * and individual contributors
 * by the @authors tag. See the copyright.txt in the distribution for a
 * full listing of individual contributors.
 *
 * This is free software; you can redistribute it and/or modify it
 * under the terms of the GNU Lesser General Public License as
 * published by the Free Software Foundation; either version 2.1 of
 * the License, or (at your option) any later version.
 *
 * This software is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public
 * License along with this software; if not, write to the Free
 * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
 * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 */

package org.mobicents.smsc.smpp;

import java.io.Serializable;

/**
 * 
 * @author sergey vetyutnev
 *
 */
public class DbSmsRoutingRule implements Serializable {

    private static final long serialVersionUID = -4693280752311155768L;

    private int id;
    private String address;
    private String systemId;

    public DbSmsRoutingRule() {
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public String getSystemId() {
        return systemId;
    }

    public void setSystemId(String systemId) {
        this.systemId = systemId;
    }

    @Override
    public String toString() {
        StringBuilder sb = new StringBuilder();

        sb.append("DbSmsRoutingRule [id=");
        sb.append(id);
        sb.append(", address=");
        sb.append(address);
        sb.append(", systemId=");
        sb.append(systemId);

        sb.append("]");

        return sb.toString();
    }
}
